fix: use fixed EigenLayer interval for timestamp alignment
The `calculate_aligned_start_timestamp` function was using a configurable
`duration` parameter for calculating period boundaries. This breaks if
`RewardsDuration` is changed via governance, as historical period boundaries
would be calculated incorrectly.

EigenLayer's RewardsCoordinator has a hardcoded `CALCULATION_INTERVAL_SECONDS`
of 86400 (1 day) that `startTimestamp` must be aligned to. This is independent
of the reward `duration` in the submission.

Changes:
- Add `EIGENLAYER_CALCULATION_INTERVAL` constant (86400)
- Remove `duration` parameter from `calculate_aligned_start_timestamp`
- Always align to EigenLayer's fixed 86400-second interval
- Update tests to reflect the new behavior

The `RewardsDuration` runtime parameter is still used in the ABI-encoded
calldata to tell EigenLayer how long to distribute rewards, but alignment
is now always to day boundaries.